more idiomatic if-statements in pattern matches
this implements the suggestions given in https://github.com/isabelle-prover/isabelle-go-codegen/issues/5.

To remove extra {} blocks in the translation of multi-equation functions,
this introduces a new reordering of patterns: we first check patterns which
actually deconstruct something (i.e. which are an IConst or application),
and only then apply the renamings for patterns which consist only of an IVar.

This ensures that even without the {} block we won't "leak" variables into
the scope of other function equations, since we're guaranteed to already be
within at least one if-statement, which is itself its own scope.